Renames system_util to swri_system_util. Refs #231.
