+++ swri_system_util/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(system_util)
+project(swri_system_util)
 
 find_package(catkin REQUIRED COMPONENTS roscpp)
