+++ swri_system_util/mainpage.dox
@@ -2,7 +2,7 @@
 \mainpage
 \htmlinclude manifest.html
 
-\b system_util 
+\b swri_system_util 
 
 <!-- 
 Provide an overview of your package.
