--- system_util/launch/file_util.test
@@ -1,7 +1,0 @@
-<launch>
-  <param name="path1" value="$(find system_util)" />
-  <param name="path2" value="$(find system_util)/src" />
-  <param name="path3" value="$(find system_util)/include/system_util" />
-  
-  <test test-name="test_file_util" pkg="system_util" type="test_file_util" />
-</launch>
